+++ docs/en/changes/changes.md
@@ -57,6 +57,8 @@
 * Support HTTP URI formatting through `ai-pipeline` to do pattern recognition.
 * Add new HTTP URI grouping engine with benchmark. 
 * [Breaking Change] Use the new HTTP URI grouping engine to replace the old regex based mechanism.
+* Support `sumLabeled` in `MAL`.
+* Migrate from kubernetes-client/java to fabric8 client.
 * support `sumLabeled` in `MAL`.
 * Envoy ALS generated relation metrics considers http status codes >= 400 has an error at the client side.